% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_process.R
\name{inhomogeneous_process}
\alias{inhomogeneous_process}
\title{Simulate an Inhomogeneous Poisson Process}
\usage{
inhomogeneous_process(N0, intensity)
}
\description{
We just thin out an ordinary poisson process.
}
