% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_process.R
\name{matern_process}
\alias{matern_process}
\title{Functions for Simulating Image Parameters
Simulate from a Matern Process}
\usage{
matern_process(x, nu = 1, alpha = 1)
}
\description{
Functions for Simulating Image Parameters
Simulate from a Matern Process
}
\examples{
x <- expand.grid(seq(0.1, 1, 0.05), seq(0.1, 1, 0.05))
process_df <- matern_process(x, 1, 1)
ggplot(process_df) +
 geom_tile(aes(x = Var1, y = Var2, fill = z))
}
