% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_process.R
\name{relative_intensities}
\alias{relative_intensities}
\title{Simulate Matern Probabilities}
\usage{
relative_intensities(x, K = 4, betas = NULL, ...)
}
\description{
Simulate Matern Probabilities
}
\details{
#' See page 551 in SPATIAL AND SPATIO-TEMPORAL LOG-GAUSSIAN COX
PROCESSES
}
\examples{
library("reshape2")
x <- expand.grid(seq(0.1, 1, 0.05), seq(0.1, 1, 0.05))
probs <- relative_intensities(x, nu = 1)
}
